Updates for EditableGridPanel conversions in platform core-components page - factor loadEditorModelData() out of AssayWizardModel.ts - export EditableGridPanel and loadEditorModelData from index.ts
