+++ packages/components/releaseNotes/components.md
@@ -1,6 +1,12 @@
 # @labkey/components
 Components, models, actions, and utility functions for LabKey applications and pages.
 
+### version TBD
+*Released*: TBD
+* ...
+  * factor loadEditorModelData() out of AssayWizardModel.ts
+  * export EditableGridPanel and loadEditorModelData from index.ts
+
 ### version 2.131.0
 *Released*: 8 February 2022
 * Issue 44711: Field editor update to show confirm modal on change of data type for saved field
