+++ packages/components/src/index.ts
@@ -210,6 +210,8 @@ import { ActionMapper, URL_MAPPERS, URLResolver, URLService } from './internal/u
 import { getHelpLink, HELP_LINK_REFERRER, HelpLink, SAMPLE_ALIQUOT_TOPIC } from './internal/util/helpLinks';
 import { AssayResolver, AssayRunResolver, ListResolver, SamplesResolver } from './internal/url/AppURLResolver';
 import { QueryGridPanel } from './internal/components/QueryGridPanel';
+import { loadEditorModelData } from './internal/components/editable/utils';
+import { EditableGridPanel } from './internal/components/editable/EditableGridPanel';
 import { EditableGridPanelDeprecated } from './internal/components/editable/EditableGridPanelDeprecated';
 import { EditableGridPanelForUpdate } from './internal/components/editable/EditableGridPanelForUpdate';
 import { EditableGridLoader } from './internal/components/editable/EditableGridLoader';
@@ -821,9 +823,11 @@ export {
     invalidateQueryDetailsCache,
     invalidateQueryDetailsCacheKey,
     // editable grid related items
+    loadEditorModelData,
     MAX_EDITABLE_GRID_ROWS,
     EditableGridLoaderFromSelection,
     EditableGridLoader,
+    EditableGridPanel,
     EditableGridPanelDeprecated,
     EditableGridPanelForUpdate,
     EditableGridModal,
@@ -1410,7 +1414,7 @@ export type { MessageFunction, NotificationItemProps } from './internal/componen
 export type { VisGraphNode } from './internal/components/lineage/vis/VisGraphGenerator';
 export type { ITab } from './internal/components/navigation/SubNav';
 export type { NotificationCreatable } from './internal/components/notifications/actions';
-export type { IDataViewInfo } from './internal/models';
+export type { IDataViewInfo, EditorModelProps } from './internal/models';
 export type { HeatMapCell } from './internal/components/heatmap/HeatMap';
 export type { InjectedAssayModel, WithAssayModelProps } from './internal/components/assay/withAssayModels';
 export type { SearchResultCardData } from './internal/components/search/models';
