+++ packages/components/src/internal/components/assay/AssayWizardModel.ts
@@ -1,5 +1,5 @@
-import { fromJS, List, Map, OrderedMap, Record, Set } from 'immutable';
-import { AssayDOM, Utils } from '@labkey/api';
+import { fromJS, List, Map, OrderedMap, Record } from 'immutable';
+import { AssayDOM } from '@labkey/api';
 
 import {
     AppURL,
@@ -11,10 +11,9 @@ import {
     QueryInfo,
     QueryModel,
 } from '../../..';
-import { genCellKey, getLookupDisplayValue, getLookupValueDescriptors } from '../../actions';
 import { AssayUploadTabs } from '../../constants';
-import { ValueDescriptor } from '../../models';
 import { generateNameWithTimestamp } from '../../util/Date';
+import { loadEditorModelData } from "../editable/utils";
 
 // exported for jest testing
 export function parseDataTextToRunRows(rawData: string): any[] {
@@ -236,68 +235,14 @@ export class AssayWizardModel
         return assayData;
     }
 
-    getInitialQueryModelData(): { rows: { [key: string]: any }; orderedRows: string[] } {
+    getInitialQueryModelData(): { rows: { [key: string]: any }; orderedRows: string[], queryInfo: QueryInfo } {
         const { assayDef, selectedSamples } = this;
         const sampleColumnData = assayDef.getSampleColumn();
         const sampleColInResults = sampleColumnData && sampleColumnData.domain === AssayDomainTypes.RESULT;
         const hasSamples = sampleColInResults && selectedSamples;
         // We only care about passing samples to the data grid if there is a sample column in the results domain.
         const rows = hasSamples ? selectedSamples.toJS() : {};
-        return { rows, orderedRows: Object.keys(rows) };
-    }
-
-    // TODO: this method is essentially a duplicate of loadDataForEditor. We should probably make a version that can be
-    //  shared because we're going to need this outside of Assays when we convert more usages of
-    //  EditableGridPanelDeprecated
-    async getInitialEditorModelData(queryModelData: Partial<QueryModel>): Promise<Partial<EditorModel>> {
-        const { orderedRows, rows } = queryModelData;
-        const columns = this.queryInfo.getInsertColumns();
-        const lookupValueDescriptors = await getLookupValueDescriptors(
-            columns.toArray(),
-            fromJS(rows),
-            fromJS(orderedRows)
-        );
-        let cellValues = Map<string, List<ValueDescriptor>>();
-
-        // data is initialized in column order
-        columns.forEach((col, cn) => {
-            orderedRows.forEach((id, rn) => {
-                const row = rows[id];
-                const cellKey = genCellKey(cn, rn);
-                const value = row[col.fieldKey];
-
-                if (Array.isArray(value)) {
-                    // assume to be list of {displayValue, value} objects
-                    cellValues = cellValues.set(
-                        cellKey,
-                        value.reduce(
-                            (list, v) => list.push({ display: v.displayValue, raw: v.value }),
-                            List<ValueDescriptor>()
-                        )
-                    );
-                } else {
-                    let cellValue = List([{ display: value, raw: value }]);
-
-                    // Issue 37833: try resolving the value for the lookup to get the displayValue to show in the grid
-                    // cell
-                    if (col.isLookup() && Utils.isNumber(value)) {
-                        const descriptors = lookupValueDescriptors[col.lookupKey];
-                        if (descriptors) {
-                            cellValue = List(descriptors.filter(descriptor => descriptor.raw === value));
-                        }
-                    }
-
-                    cellValues = cellValues.set(cellKey, cellValue);
-                }
-            });
-        });
-
-        return {
-            cellValues,
-            colCount: columns.size,
-            deletedIds: Set<any>(),
-            rowCount: orderedRows.length,
-        };
+        return { rows, orderedRows: Object.keys(rows), queryInfo: this.queryInfo };
     }
 
     /**
@@ -306,7 +251,7 @@ export class AssayWizardModel
      */
     async getInitialGridData(): Promise<{ editorModel: Partial<EditorModel>; queryModel: Partial<QueryModel> }> {
         const queryModelData = this.getInitialQueryModelData();
-        const editorModelData = await this.getInitialEditorModelData(queryModelData);
+        const editorModelData = await loadEditorModelData(queryModelData);
         return { editorModel: editorModelData, queryModel: queryModelData };
     }
 }
